Add navbar products header (right component)
