+++ blocks/header/header-megamenu.js
@@ -1,40 +1,22 @@
 import handleViewportChanges from './header-events.js';
 import { fetchHeaderContent } from './header.js';
+import buildProductsMegaMenu from './menus/products.js';
+import buildMegaMenuLeftMenus from './menus/left-submenus.js';
 import { getMetadata, toClassName } from '../../scripts/lib-franklin.js';
-
-function insertMegaMenuBackgroundImage(navContent, submenuContent) {
-  const backgroundImg = navContent.querySelector('.submenu-background img');
-  submenuContent.style.backgroundImage = `url(${backgroundImg.src})`;
-}
-
-function buildMegaMenuLeftMenus(submenuContent) {
-  // get all H2s and create a list of them
-  const h2s = [...submenuContent.querySelectorAll('h2')];
-  const h2List = document.createElement('ul');
-  h2List.classList.add('menu-nav-submenu-sections');
-
-  // add H2s to list
-  h2s.forEach((h2) => {
-    const h2ListItem = document.createElement('li');
-    h2ListItem.classList.add('menu-nav-submenu-section');
-    h2ListItem.innerHTML = h2.outerHTML;
-    h2List.append(h2ListItem);
-  });
-  return h2List;
-}
 
 function buildMegaMenu(navContent, submenuContent) {
   const productsSubmenu = document.createElement('div');
   productsSubmenu.append(submenuContent.querySelector('h1'));
   productsSubmenu.append(buildMegaMenuLeftMenus(submenuContent));
   submenuContent.innerHTML = productsSubmenu.outerHTML;
-  insertMegaMenuBackgroundImage(navContent, submenuContent);
+  const backgroundImg = navContent.querySelector('.submenu-background img');
+  submenuContent.style.backgroundImage = `url(${backgroundImg.src})`;
 }
 
 function createSubmenuBuildersMap() {
   // create map of submenu name to function
   const submenus = new Map();
-  submenus.set('products', buildMegaMenu);
+  submenus.set('products', buildProductsMegaMenu);
   submenus.set('applications', buildMegaMenu);
   submenus.set('resources', buildMegaMenu);
   submenus.set('service-support', buildMegaMenu);
