+++ blocks/header/header.css
@@ -199,7 +199,7 @@ header .nav-tabs .menu-nav-submenu {
   background: var(--background-color) url('/images/header-background-filler.jpg') no-repeat right bottom;
 }
 
-header .nav-tabs .menu-nav-submenu div {
+header .nav-tabs .menu-nav-submenu > div {
   padding: 0;
   margin: 0 auto;
   display: block;
@@ -216,20 +216,26 @@ header .nav-tabs .menu-nav-submenu-sections li h2 {
   margin: 0 2px 0 0;
 }
 
-.header .nav-tabs .menu-nav-submenu > div > h1 {
+header .nav-tabs .menu-nav-submenu h1 {
   font-size: 36px;
   font-family: 'Proxima Nova Light', sans-serif;
   color: var(--header-mega-menu-font-color);
   border-bottom: 0;
   padding: 9px 20px 6px;
-  margin: 20px 2px 0 0;
+  margin: -10px 2px 0 0;
 }
 
-.header .nav-tabs .menu-nav-submenu a {
+header .menu-nav-submenu h1 > a {
   color: inherit;
+  text-decoration: none;
 }
 
-.header .nav-tabs .menu-nav-submenu-head {
+header .nav-tabs .menu-nav-submenu-sections a {
+  color: inherit;
+  text-decoration: none;
+}
+
+header .nav-tabs .menu-nav-submenu-head {
   background: transparent;
   box-shadow: none !important;
   position: absolute;
@@ -265,6 +271,144 @@ header .nav-tabs .menu-nav-submenu-sections {
   border-top: 1px solid var(--header-mega-menu-border-color);
   width: 330px;
   margin-top: 0;
+}
+
+header .nav-tabs .right-submenu {
+  background: transparent;
+  box-shadow: none !important;
+  position: absolute;
+  top: 75px;
+  transform: translateX(370px);
+  width: calc(100% - 370px) !important;
+  left: 0;
+  right: 0;
+  border: 0;
+  padding: 0;
+  overflow: hidden;
+  z-index: 0;
+  margin: 0 !important;
+}
+
+header .nav-tabs .right-submenu-row {
+  width: 100%;
+  position: static;
+  min-height: 500px;
+  margin-left: 0;
+  margin-right: 0;
+  float: left;
+}
+
+header .flex-space-between {
+  display: flex;
+  justify-content: space-between;
+}
+
+header .large-card {
+  background-size: 250px;
+  color: var(--background-color);
+  background-position: 110% 154%;
+  background-color: var(--header-mega-menu-product-cards-background-color);
+  background-repeat: no-repeat;
+  position: relative;
+  padding-top: 20px;
+  padding-right: 15px;
+  padding-left: 15px;
+  margin-right: 20px;
+  width: 60%;
+  float: left;
+  height: 470px;
+}
+
+header .large-card ul {
+  padding: 0;
+  width: 45%;
+  position: relative;
+  min-height: 1px;
+  padding-left: 20px;
+  margin-bottom: 40px;
+  font-family: "Quicksand",sans-serif;
+  font-weight: 400;
+  font-size: 14px;
+  text-align: left;
+  list-style-type: circle;
+}
+
+header .large-card ul > li {
+  display: list-item;
+  position: relative;
+  float: left;
+  border: 0;
+  position: static;
+  min-width: inherit !important;
+  margin: 0;
+  padding: 0;
+  box-shadow: none;
+  margin-bottom: 8px;
+  color: var(--background-color);
+}
+
+header .large-card h3 {
+  border: 0;
+  color: var(--background-color);
+  padding: 0;
+  padding-bottom: 10px;
+  font-size: 18px;
+  font-family: "Proxima Nova Regular", sans-serif;
+  margin-top: 0;
+  font-weight: 300;
+  margin-bottom: 10px;
+}
+
+header .large-card h3 a {
+  color: var(--header-mega-menu-green);
+  text-decoration: none;
+  text-transform: initial;
+}
+
+header .large-card img {
+  right: 0;
+  width: 50%;
+  max-height: 250px;
+  object-fit: contain;
+  display: block;
+  margin: auto;
+  transition: all .2s;
+}
+
+header .image-right {
+  background-position: 110% 154%;
+}
+
+header .image-right img {
+  position: absolute; 
+  bottom: -60px;
+}
+
+header .image-center  {
+  background-position: 50% 390%;
+  background-size: 80%;
+}
+
+header .image-center img {
+  transform: scale(1);
+  position: relative;
+  width: 185px;
+  bottom: -100px;
+  min-height: 200px;
+  margin-left: 30px;
+}
+
+header .axis-robotic-arm {
+  background-image: url('../../images/header-menus/BAB400-6-axis-robotic-arm-bg.png');
+}
+
+header .dispencell {
+  background-image: url('../../images/header-menus/dispencell-single-cell-dispenser-bg.png');
+}
+
+header .large-card img:hover {
+  cursor: pointer;
+  transform: scale(1.2);
 }
 
 @media (min-width: 768) {
@@ -371,7 +515,7 @@ header .nav-tabs .menu-nav-submenu-sections {
     line-height: 18px;
   }
 
-  header .nav-tabs li a {
+  header .nav-tabs li > a {
     padding: 11px 0;
     text-transform: uppercase;
     text-align: left;
@@ -379,6 +523,8 @@ header .nav-tabs .menu-nav-submenu-sections {
     margin-bottom: -1px;
     border: 0;
     margin-right: 0;
+    color: inherit;
+    text-decoration: none;
   }
 
   header .nav-tabs > li.searchlink>a {
@@ -449,7 +595,7 @@ header .nav-tabs .menu-nav-submenu-sections {
     display: initial;
   }
 
-  header .nav-tabs .menu-nav-submenu div {
+  header .nav-tabs .menu-nav-submenu > div {
     min-height: auto !important;
     width: 970px;
   }
@@ -473,6 +619,23 @@ header .nav-tabs .menu-nav-submenu-sections {
 @media (min-width: 1200px) {
   header .container {
     width: 1170px;
+  }
+
+  header .large-card {
+    height: 360px;
+  }
+
+  header .dispencell img {
+    bottom: -30px;
+  }
+
+  header .axis-robotic-arm img  {
+    bottom: -10px;
+  }
+
+  header .image-center img {
+    bottom: -10px;
+    margin-left: 70px;
   }
 
   header .nav-tabs .menu-nav-category {
